dexter/src/serializers/promptsserializer.py
from rest_framework import serializers
from src.models.prompts import Prompt

class PromptSerializer(serializers.ModelSerializer):
    class Meta:
        model = Prompt
        fields = "__all__"
