dexter/requirements.txt
langchain
google-search-results
openai
websockets
django
djangorestframework
django-cors-headers
djangorestframework_simplejwt
